Seperate the Task classes into their respective files
