+++ src/main/java/sisyphus/TodoTask.java
@@ -1,0 +1,12 @@
+package sisyphus;
+
+public class TodoTask extends Task {
+
+    public TodoTask(String name) {
+        super(name);
+    }
+
+    public String toString() {
+        return "[T] " + super.toString();
+    }
+}
